implemented an abstract K2Raster class Ravenjoin now uses this abstract class instead of K2Raster This allows us to have both a long and int version of K2Raster with minimal code duplication
